+++ NEWS.md
@@ -15,6 +15,7 @@
 * Performance has been improved for `fcm()` and for `textmodel_wordfish()`.
 * `dfm()` now correctly passes through all `...` arguments to `tokens()`.  (#1121)
 * All `dfm_*()` functions now work correctly with empty dfm objects.  (#1133)
+* Fixed a bug in `dfm_weight()` for named weight vectors (#1150)
 
 
 ### Behaviour Changes